finished endpoints and functions for deleting and adding recipes

+++ db/remove_from_favorites.sql
@@ -1,2 +1,3 @@
 delete from favorites
-where id=$1;
+where recipe_id=$1
+and user_id=$2;
